+++ HunterPie.Integrations/Datasources/MonsterHunterRise/Definitions/MHRQurioPartStructure.cs
@@ -6,6 +6,6 @@ namespace HunterPie.Integrations.Datasources.MonsterHunterRise.Definitions;
 public struct MHRQurioPartStructure
 {
     [FieldOffset(0x10)] public bool IsActive;
-    [FieldOffset(0x18)] public long HealthPtr;
+    [FieldOffset(0x18)] public nint HealthPtr;
     [FieldOffset(0x38)] public float MaxHealth;
 }
